修复sql

+++ app/src/main/java/com/cwd/money/mysql/table/DeltaInfo.kt
@@ -1,8 +1,8 @@
 package com.cwd.money.mysql.table
 
 data class DeltaInfo(val count:Int,
-                     val turn_ten:Int,
-                     val turn_twn:Int,
-                     val turn_thirty:Int,
-                     val turn_ot:Int,
+                     val turn_ten:Float,
+                     val turn_twn:Float,
+                     val turn_thirty:Float,
+                     val turn_ot:Float,
                      val shareInfo: ShareInfo)
